Refactor GaussianCNNBaseline (#1664)
