+++ examples/tf/ppo_memorize_digits.py
@@ -45,15 +45,15 @@ def ppo_memorize_digits(ctxt=None, seed=1, batch_size=4000):
 
         baseline = GaussianCNNBaseline(
             env_spec=env.spec,
-            regressor_args=dict(filters=(
-                                            (32, (5, 5)),
-                                            (64, (3, 3)),
-                                            (64, (2, 2)),
-                                        ),
-                                strides=(4, 2, 1),
-                                padding='VALID',
-                                hidden_sizes=(256, ),
-                                use_trust_region=True))  # yapf: disable
+            filters=(
+                (32, (5, 5)),
+                (64, (3, 3)),
+                (64, (2, 2)),
+            ),
+            strides=(4, 2, 1),
+            padding='VALID',
+            hidden_sizes=(256, ),
+            use_trust_region=True)  # yapf: disable
 
         algo = PPO(env_spec=env.spec,
                    policy=policy,
