+++ src/garage/np/baselines/baseline.py
@@ -3,33 +3,7 @@ import abc
 
 
 class Baseline(abc.ABC):
-    """Base class for all baselines.
-
-    Args:
-        env_spec (garage.envs.env_spec.EnvSpec): Environment specification.
-
-    """
-
-    def __init__(self, env_spec):
-        self._mdp_spec = env_spec
-
-    @abc.abstractmethod
-    def get_param_values(self):
-        """Get parameter values.
-
-        Returns:
-            List[np.ndarray]: A list of values of each parameter.
-
-        """
-
-    @abc.abstractmethod
-    def set_param_values(self, flattened_params):
-        """Set param values.
-
-        Args:
-            flattened_params (np.ndarray): A numpy array of parameter values.
-
-        """
+    """Base class for all baselines."""
 
     @abc.abstractmethod
     def fit(self, paths):
@@ -41,21 +15,13 @@ class Baseline(abc.ABC):
         """
 
     @abc.abstractmethod
-    def predict(self, path):
+    def predict(self, paths):
         """Predict value based on paths.
 
         Args:
-            path (dict[numpy.ndarray]): Sample paths.
+            paths (dict[numpy.ndarray]): Sample paths.
 
         Returns:
             numpy.ndarray: Predicted value.
-
-        """
-
-    def log_diagnostics(self, paths):
-        """Log diagnostic information.
-
-        Args:
-            paths (list[dict]): A list of collected paths.
 
         """
