+++ src/garage/tf/algos/reps.py
@@ -198,20 +198,9 @@ class REPS(RLAlgorithm):  # noqa: D416
 
         samples_data['average_return'] = np.mean(undiscounted_returns)
 
-        self.log_diagnostics(samples_data)
         logger.log('Optimizing policy...')
         self.optimize_policy(samples_data)
         return samples_data['average_return']
-
-    def log_diagnostics(self, paths):
-        """Log diagnostic information.
-
-        Args:
-            paths (list[dict]): A list of collected paths.
-
-        """
-        logger.log('Logging diagnostics...')
-        self._baseline.log_diagnostics(paths)
 
     def __getstate__(self):
         """Parameters to save in snapshot.
