+++ src/garage/tf/regressors/__init__.py
@@ -4,7 +4,6 @@ from garage.tf.regressors.categorical_mlp_regressor import (
     CategoricalMLPRegressor)
 from garage.tf.regressors.continuous_mlp_regressor import (
     ContinuousMLPRegressor)
-from garage.tf.regressors.gaussian_cnn_regressor import GaussianCNNRegressor
 from garage.tf.regressors.gaussian_cnn_regressor_model import (
     GaussianCNNRegressorModel)
 from garage.tf.regressors.gaussian_mlp_regressor import GaussianMLPRegressor
@@ -12,7 +11,6 @@ from garage.tf.regressors.regressor import Regressor, StochasticRegressor
 
 __all__ = [
     'BernoulliMLPRegressor', 'CategoricalMLPRegressor',
-    'ContinuousMLPRegressor', 'GaussianCNNRegressor',
-    'GaussianCNNRegressorModel', 'GaussianMLPRegressor', 'Regressor',
-    'StochasticRegressor'
+    'ContinuousMLPRegressor', 'GaussianCNNRegressorModel',
+    'GaussianMLPRegressor', 'Regressor', 'StochasticRegressor'
 ]
