+++ tests/garage/tf/algos/test_trpo.py
@@ -153,13 +153,13 @@ class TestTRPOCNNCubeCrash(TfGraphTestCase):
                                           padding='VALID',
                                           hidden_sizes=(32, 32))
 
-            baseline = GaussianCNNBaseline(
-                env_spec=env.spec,
-                regressor_args=dict(filters=((32, (8, 8)), (64, (4, 4))),
-                                    strides=(4, 2),
-                                    padding='VALID',
-                                    hidden_sizes=(32, 32),
-                                    use_trust_region=True))
+            baseline = GaussianCNNBaseline(env_spec=env.spec,
+                                           filters=((32, (8, 8)), (64, (4,
+                                                                        4))),
+                                           strides=(4, 2),
+                                           padding='VALID',
+                                           hidden_sizes=(32, 32),
+                                           use_trust_region=True)
 
             algo = TRPO(env_spec=env.spec,
                         policy=policy,
